chore (model) : `domain` 패키지 명 변경
- domain -> model 패키지 명 변경
