--- src/main/java/baseball/domain/Player.java
@@ -1,5 +1,0 @@
-package baseball.domain;
-
-interface Player {
-
-}
