--- src/main/java/baseball/domain/concept/Concept.java
@@ -1,8 +1,0 @@
-package baseball.domain.concept;
-
-import java.util.Collection;
-
-public interface Concept {
-
-    Collection<Position> thinkPositions();
-}
