+++ src/main/java/baseball/model/Player.java
@@ -1,0 +1,5 @@
+package baseball.model;
+
+interface Player {
+
+}
