+++ src/main/java/baseball/model/concept/Concept.java
@@ -1,0 +1,8 @@
+package baseball.model.concept;
+
+import java.util.Collection;
+
+public interface Concept {
+
+    Collection<Position> thinkPositions();
+}
